refactor: change case from IgnorePHPUnitWarnings to IgnorePhpunitWarnings
