+++ src/Metadata/Parser/AttributeParser.php
@@ -52,7 +52,7 @@ use PHPUnit\Framework\Attributes\ExcludeStaticPropertyFromBackup;
 use PHPUnit\Framework\Attributes\Group;
 use PHPUnit\Framework\Attributes\IgnoreDeprecations;
 use PHPUnit\Framework\Attributes\IgnorePhpunitDeprecations;
-use PHPUnit\Framework\Attributes\IgnorePHPUnitWarnings;
+use PHPUnit\Framework\Attributes\IgnorePhpunitWarnings;
 use PHPUnit\Framework\Attributes\Large;
 use PHPUnit\Framework\Attributes\Medium;
 use PHPUnit\Framework\Attributes\PostCondition;
@@ -885,10 +885,10 @@ final readonly class AttributeParser implements Parser
 
                     break;
 
-                case IgnorePHPUnitWarnings::class:
-                    assert($attributeInstance instanceof IgnorePHPUnitWarnings);
+                case IgnorePhpunitWarnings::class:
+                    assert($attributeInstance instanceof IgnorePhpunitWarnings);
 
-                    $result[] = Metadata::ignorePHPUnitWarnings($attributeInstance->messagePattern());
+                    $result[] = Metadata::ignorePhpunitWarnings($attributeInstance->messagePattern());
 
                     break;
             }
