+++ tests/unit/Metadata/MetadataTest.php
@@ -52,6 +52,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -80,7 +81,6 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isUsesMethod());
         $this->assertFalse($metadata->isUsesTrait());
         $this->assertFalse($metadata->isWithoutErrorHandler());
-        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
 
         $this->assertTrue($metadata->isMethodLevel());
         $this->assertFalse($metadata->isClassLevel());
@@ -114,6 +114,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -142,7 +143,6 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isUsesMethod());
         $this->assertFalse($metadata->isUsesTrait());
         $this->assertFalse($metadata->isWithoutErrorHandler());
-        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
 
         $this->assertTrue($metadata->isMethodLevel());
         $this->assertFalse($metadata->isClassLevel());
@@ -176,6 +176,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -204,7 +205,6 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isUsesMethod());
         $this->assertFalse($metadata->isUsesTrait());
         $this->assertFalse($metadata->isWithoutErrorHandler());
-        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
 
         $this->assertFalse($metadata->enabled());
         $this->assertTrue($metadata->isClassLevel());
@@ -239,6 +239,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -267,7 +268,6 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isUsesMethod());
         $this->assertFalse($metadata->isUsesTrait());
         $this->assertFalse($metadata->isWithoutErrorHandler());
-        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
 
         $this->assertFalse($metadata->enabled());
         $this->assertTrue($metadata->isMethodLevel());
@@ -302,6 +302,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -330,7 +331,6 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isUsesMethod());
         $this->assertFalse($metadata->isUsesTrait());
         $this->assertFalse($metadata->isWithoutErrorHandler());
-        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
 
         $this->assertFalse($metadata->enabled());
         $this->assertTrue($metadata->isClassLevel());
@@ -365,6 +365,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -393,7 +394,6 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isUsesMethod());
         $this->assertFalse($metadata->isUsesTrait());
         $this->assertFalse($metadata->isWithoutErrorHandler());
-        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
 
         $this->assertFalse($metadata->enabled());
         $this->assertTrue($metadata->isMethodLevel());
@@ -428,6 +428,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -456,7 +457,6 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isUsesMethod());
         $this->assertFalse($metadata->isUsesTrait());
         $this->assertFalse($metadata->isWithoutErrorHandler());
-        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
 
         $this->assertTrue($metadata->isMethodLevel());
         $this->assertFalse($metadata->isClassLevel());
@@ -489,6 +489,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -517,7 +518,6 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isUsesMethod());
         $this->assertFalse($metadata->isUsesTrait());
         $this->assertFalse($metadata->isWithoutErrorHandler());
-        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
 
         $this->assertTrue($metadata->isMethodLevel());
         $this->assertFalse($metadata->isClassLevel());
@@ -551,6 +551,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -579,7 +580,6 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isUsesMethod());
         $this->assertFalse($metadata->isUsesTrait());
         $this->assertFalse($metadata->isWithoutErrorHandler());
-        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
 
         $this->assertSame(self::class, $metadata->className());
 
@@ -617,6 +617,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -645,7 +646,6 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isUsesMethod());
         $this->assertFalse($metadata->isUsesTrait());
         $this->assertFalse($metadata->isWithoutErrorHandler());
-        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
 
         $this->assertSame($namespace, $metadata->namespace());
 
@@ -681,6 +681,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -709,7 +710,6 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isUsesMethod());
         $this->assertFalse($metadata->isUsesTrait());
         $this->assertFalse($metadata->isWithoutErrorHandler());
-        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
 
         $this->assertSame(self::class, $metadata->className());
 
@@ -745,6 +745,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -773,7 +774,6 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isUsesMethod());
         $this->assertFalse($metadata->isUsesTrait());
         $this->assertFalse($metadata->isWithoutErrorHandler());
-        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
 
         $this->assertSame(self::class, $metadata->interfaceName());
 
@@ -809,6 +809,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -837,7 +838,6 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isUsesMethod());
         $this->assertFalse($metadata->isUsesTrait());
         $this->assertFalse($metadata->isWithoutErrorHandler());
-        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
 
         $this->assertSame('f', $metadata->functionName());
 
@@ -873,6 +873,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -901,7 +902,6 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isUsesMethod());
         $this->assertFalse($metadata->isUsesTrait());
         $this->assertFalse($metadata->isWithoutErrorHandler());
-        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
 
         $this->assertSame(self::class, $metadata->className());
         $this->assertSame('testCanBeCoversMethod', $metadata->methodName());
@@ -938,6 +938,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -966,7 +967,6 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isUsesMethod());
         $this->assertFalse($metadata->isUsesTrait());
         $this->assertFalse($metadata->isWithoutErrorHandler());
-        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
 
         $this->assertTrue($metadata->isMethodLevel());
         $this->assertFalse($metadata->isClassLevel());
@@ -1000,6 +1000,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -1028,7 +1029,6 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isUsesMethod());
         $this->assertFalse($metadata->isUsesTrait());
         $this->assertFalse($metadata->isWithoutErrorHandler());
-        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
 
         $this->assertTrue($metadata->isClassLevel());
         $this->assertFalse($metadata->isMethodLevel());
@@ -1062,6 +1062,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -1090,7 +1091,6 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isUsesMethod());
         $this->assertFalse($metadata->isUsesTrait());
         $this->assertFalse($metadata->isWithoutErrorHandler());
-        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
 
         $this->assertSame(ExampleTrait::class, $metadata->traitName());
 
@@ -1126,6 +1126,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -1154,7 +1155,6 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isUsesMethod());
         $this->assertFalse($metadata->isUsesTrait());
         $this->assertFalse($metadata->isWithoutErrorHandler());
-        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
 
         $this->assertSame(self::class, $metadata->className());
         $this->assertSame('method', $metadata->methodName());
@@ -1191,6 +1191,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -1219,7 +1220,6 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isUsesMethod());
         $this->assertFalse($metadata->isUsesTrait());
         $this->assertFalse($metadata->isWithoutErrorHandler());
-        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
 
         $this->assertSame(self::class, $metadata->className());
         $this->assertFalse($metadata->deepClone());
@@ -1257,6 +1257,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -1285,7 +1286,6 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isUsesMethod());
         $this->assertFalse($metadata->isUsesTrait());
         $this->assertFalse($metadata->isWithoutErrorHandler());
-        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
 
         $this->assertSame(self::class, $metadata->className());
         $this->assertSame('method', $metadata->methodName());
@@ -1324,6 +1324,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -1352,7 +1353,6 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isUsesMethod());
         $this->assertFalse($metadata->isUsesTrait());
         $this->assertFalse($metadata->isWithoutErrorHandler());
-        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
 
         $this->assertTrue($metadata->isClassLevel());
         $this->assertFalse($metadata->isMethodLevel());
@@ -1386,6 +1386,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -1447,6 +1448,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -1508,6 +1510,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -1570,6 +1573,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -1632,6 +1636,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -1696,6 +1701,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -1760,6 +1766,7 @@ final class MetadataTest extends TestCase
         $this->assertTrue($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -1823,6 +1830,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertTrue($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -1884,6 +1892,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertTrue($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -1945,6 +1954,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertTrue($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -2006,6 +2016,69 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertTrue($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
+        $this->assertFalse($metadata->isRunClassInSeparateProcess());
+        $this->assertFalse($metadata->isRunInSeparateProcess());
+        $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
+        $this->assertFalse($metadata->isTest());
+        $this->assertFalse($metadata->isPreCondition());
+        $this->assertFalse($metadata->isPostCondition());
+        $this->assertFalse($metadata->isPreserveGlobalState());
+        $this->assertFalse($metadata->isRequiresMethod());
+        $this->assertFalse($metadata->isRequiresFunction());
+        $this->assertFalse($metadata->isRequiresOperatingSystem());
+        $this->assertFalse($metadata->isRequiresOperatingSystemFamily());
+        $this->assertFalse($metadata->isRequiresPhp());
+        $this->assertFalse($metadata->isRequiresPhpExtension());
+        $this->assertFalse($metadata->isRequiresPhpunit());
+        $this->assertFalse($metadata->isRequiresPhpunitExtension());
+        $this->assertFalse($metadata->isRequiresEnvironmentVariable());
+        $this->assertFalse($metadata->isWithEnvironmentVariable());
+        $this->assertFalse($metadata->isRequiresSetting());
+        $this->assertFalse($metadata->isTestDox());
+        $this->assertFalse($metadata->isTestWith());
+        $this->assertFalse($metadata->isUsesNamespace());
+        $this->assertFalse($metadata->isUsesClass());
+        $this->assertFalse($metadata->isUsesClassesThatExtendClass());
+        $this->assertFalse($metadata->isUsesClassesThatImplementInterface());
+        $this->assertFalse($metadata->isUsesFunction());
+        $this->assertFalse($metadata->isUsesMethod());
+        $this->assertFalse($metadata->isUsesTrait());
+        $this->assertFalse($metadata->isWithoutErrorHandler());
+
+        $this->assertTrue($metadata->isMethodLevel());
+        $this->assertFalse($metadata->isClassLevel());
+    }
+
+    public function testCanBeIgnorePHPUnitWarnings(): void
+    {
+        $metadata = Metadata::ignorePHPUnitWarnings(null);
+
+        $this->assertFalse($metadata->isAfter());
+        $this->assertFalse($metadata->isAfterClass());
+        $this->assertFalse($metadata->isBackupGlobals());
+        $this->assertFalse($metadata->isBackupStaticProperties());
+        $this->assertFalse($metadata->isBeforeClass());
+        $this->assertFalse($metadata->isBefore());
+        $this->assertFalse($metadata->isCoversNamespace());
+        $this->assertFalse($metadata->isCoversClass());
+        $this->assertFalse($metadata->isCoversClassesThatExtendClass());
+        $this->assertFalse($metadata->isCoversClassesThatImplementInterface());
+        $this->assertFalse($metadata->isCoversFunction());
+        $this->assertFalse($metadata->isCoversMethod());
+        $this->assertFalse($metadata->isCoversNothing());
+        $this->assertFalse($metadata->isCoversTrait());
+        $this->assertFalse($metadata->isDataProvider());
+        $this->assertFalse($metadata->isDependsOnClass());
+        $this->assertFalse($metadata->isDependsOnMethod());
+        $this->assertFalse($metadata->isDisableReturnValueGenerationForTestDoubles());
+        $this->assertFalse($metadata->isDoesNotPerformAssertions());
+        $this->assertFalse($metadata->isExcludeGlobalVariableFromBackup());
+        $this->assertFalse($metadata->isExcludeStaticPropertyFromBackup());
+        $this->assertFalse($metadata->isGroup());
+        $this->assertFalse($metadata->isIgnoreDeprecations());
+        $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertTrue($metadata->isIgnorePhpunitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -2067,6 +2140,7 @@ final class MetadataTest extends TestCase
         $this->assertTrue($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -2130,6 +2204,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertTrue($metadata->isRunTestsInSeparateProcesses());
@@ -2191,6 +2266,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertTrue($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -2252,6 +2328,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertTrue($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -2313,6 +2390,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -2374,6 +2452,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -2435,6 +2514,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -2492,6 +2572,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -2551,6 +2632,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -2614,6 +2696,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -2678,6 +2761,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -2742,6 +2826,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -2805,6 +2890,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -2868,6 +2954,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -2931,6 +3018,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -2994,6 +3082,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -3057,6 +3146,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -3125,6 +3215,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -3193,6 +3284,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -3256,6 +3348,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -3329,6 +3422,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -3394,6 +3488,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -3467,6 +3562,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -3537,6 +3633,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -3600,6 +3697,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -3668,6 +3766,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -3731,6 +3830,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -3794,6 +3894,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -3858,6 +3959,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -3922,6 +4024,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -3986,6 +4089,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -4050,6 +4154,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -4114,6 +4219,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -4178,6 +4284,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -4241,6 +4348,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -4304,6 +4412,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -4370,6 +4479,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -4433,6 +4543,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -4496,6 +4607,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -4559,6 +4671,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -4622,6 +4735,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -4684,6 +4798,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -4748,6 +4863,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
@@ -4783,68 +4899,6 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isMethodLevel());
     }
 
-    public function testCanBeIgnorePHPUnitWarnings(): void
-    {
-        $metadata = Metadata::ignorePHPUnitWarnings(null);
-
-        $this->assertFalse($metadata->isAfter());
-        $this->assertFalse($metadata->isAfterClass());
-        $this->assertFalse($metadata->isBackupGlobals());
-        $this->assertFalse($metadata->isBackupStaticProperties());
-        $this->assertFalse($metadata->isBeforeClass());
-        $this->assertFalse($metadata->isBefore());
-        $this->assertFalse($metadata->isCoversNamespace());
-        $this->assertFalse($metadata->isCoversClass());
-        $this->assertFalse($metadata->isCoversClassesThatExtendClass());
-        $this->assertFalse($metadata->isCoversClassesThatImplementInterface());
-        $this->assertFalse($metadata->isCoversFunction());
-        $this->assertFalse($metadata->isCoversMethod());
-        $this->assertFalse($metadata->isCoversNothing());
-        $this->assertFalse($metadata->isCoversTrait());
-        $this->assertFalse($metadata->isDataProvider());
-        $this->assertFalse($metadata->isDependsOnClass());
-        $this->assertFalse($metadata->isDependsOnMethod());
-        $this->assertFalse($metadata->isDisableReturnValueGenerationForTestDoubles());
-        $this->assertFalse($metadata->isDoesNotPerformAssertions());
-        $this->assertFalse($metadata->isExcludeGlobalVariableFromBackup());
-        $this->assertFalse($metadata->isExcludeStaticPropertyFromBackup());
-        $this->assertFalse($metadata->isGroup());
-        $this->assertFalse($metadata->isIgnoreDeprecations());
-        $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
-        $this->assertFalse($metadata->isRunClassInSeparateProcess());
-        $this->assertFalse($metadata->isRunInSeparateProcess());
-        $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
-        $this->assertFalse($metadata->isTest());
-        $this->assertFalse($metadata->isPreCondition());
-        $this->assertFalse($metadata->isPostCondition());
-        $this->assertFalse($metadata->isPreserveGlobalState());
-        $this->assertFalse($metadata->isRequiresMethod());
-        $this->assertFalse($metadata->isRequiresFunction());
-        $this->assertFalse($metadata->isRequiresOperatingSystem());
-        $this->assertFalse($metadata->isRequiresOperatingSystemFamily());
-        $this->assertFalse($metadata->isRequiresPhp());
-        $this->assertFalse($metadata->isRequiresPhpExtension());
-        $this->assertFalse($metadata->isRequiresPhpunit());
-        $this->assertFalse($metadata->isRequiresPhpunitExtension());
-        $this->assertFalse($metadata->isRequiresEnvironmentVariable());
-        $this->assertFalse($metadata->isWithEnvironmentVariable());
-        $this->assertFalse($metadata->isRequiresSetting());
-        $this->assertFalse($metadata->isTestDox());
-        $this->assertFalse($metadata->isTestWith());
-        $this->assertFalse($metadata->isUsesNamespace());
-        $this->assertFalse($metadata->isUsesClass());
-        $this->assertFalse($metadata->isUsesClassesThatExtendClass());
-        $this->assertFalse($metadata->isUsesClassesThatImplementInterface());
-        $this->assertFalse($metadata->isUsesFunction());
-        $this->assertFalse($metadata->isUsesMethod());
-        $this->assertFalse($metadata->isUsesTrait());
-        $this->assertFalse($metadata->isWithoutErrorHandler());
-        $this->assertTrue($metadata->isIgnorePHPUnitWarnings());
-
-        $this->assertTrue($metadata->isMethodLevel());
-        $this->assertFalse($metadata->isClassLevel());
-    }
-
     public function testCanBeWithoutErrorHandler(): void
     {
         $metadata = Metadata::withoutErrorHandler();
@@ -4873,6 +4927,7 @@ final class MetadataTest extends TestCase
         $this->assertFalse($metadata->isGroup());
         $this->assertFalse($metadata->isIgnoreDeprecations());
         $this->assertFalse($metadata->isIgnorePhpunitDeprecations());
+        $this->assertFalse($metadata->isIgnorePHPUnitWarnings());
         $this->assertFalse($metadata->isRunClassInSeparateProcess());
         $this->assertFalse($metadata->isRunInSeparateProcess());
         $this->assertFalse($metadata->isRunTestsInSeparateProcesses());
